Fix type and documentation of sched::thread::id()
After commit 5c652796fe8d08b7ecc87739025492505c9c1b5a, our thread id
is 32 bit, not 64 bit. This patch fixes the return type of the id()
function, and replaces an incorrect comment (which claimed that the
ids are unique during the lifetime of the system) with a more accurate
description on this function, using a doxygen comment (so it can be
seen in our doxygen API documentation).

+++ core/sched.cc
@@ -901,7 +901,7 @@ void thread::timer_fired()
     wake();
 }
 
-unsigned long thread::id()
+unsigned int thread::id()
 {
     return _id;
 }
